fix(posit): make manipulators.hpp actually independent of the stream layer (review on #1390)
The review is right, and it caught the split's central claim being false.

manipulators.hpp is meant to be the <iomanip> half -- posit to std::string, no
streams. It was not: pretty_print and info_print stream the posit itself
(str << ... << p), which needs operator<< from iostream.hpp. So core.hpp plus
manipulators.hpp, without iostream.hpp, did not compile. The <iomanip> layer
secretly depended on the <iostream> layer, which inverts the split this PR argues
for.

Reproducing it surfaced a second break, also mine: quadrant() calls the free
sign(p) from attributes.hpp, and manipulators.hpp never included it. That worked
only through the posit.hpp umbrella, because ADL defers the lookup to
instantiation -- by which point the umbrella has included everything. The facet
headers were never self-contained; the core.hpp claim is what exposed it.

Both fixed:
- pretty_print and info_print move to iostream.hpp. They stream, so they belong
  with the streams.
- manipulators.hpp includes core.hpp and attributes.hpp, so it stands on its own
  rather than depending on the umbrella's include order.

AND THIS TIME THE CLAIM IS TESTED RATHER THAN ASSERTED. core.hpp plus
manipulators.hpp now compiles AND RUNS on its own, and the only <iostream> in that
graph enters through core.hpp -> integer.hpp, the pre-existing dependency this PR
already documents. manipulators.hpp pulls <sstream>, which it genuinely needs to
build strings, and no <iostream>.

Worth recording: this is the third structural claim in this PR that did not
survive checking -- the posit_io.hpp naming, an include sweep whose pattern
matched \bstd\b and so reported nothing, and now the layer independence. The
numeric claims in this epic have been measured throughout; the structural ones I
was asserting.

Verified: 38/38 umbrellas on gcc 13.3, clang 18.1 and riscv64-linux-gnu cross; the
8 test files using pretty_print/info_print still compile; CI_LITE builds and passes
456/456; check-ascii clean; zero added lines over 120 columns.

Relates to #1334
